+++ Arrete.h
@@ -6,7 +6,7 @@
 class Arrete {
 private:
     int m_num;
-    int m_tps;
+    float m_tps;
     std::string m_type;
     std::string m_nom;
 public:
@@ -16,12 +16,12 @@ public:
     int getNum()const;
     std::string getType()const;
     std::string getNom()const;
-    int getTps()const;
+    float getTps()const;
     ///Acesseur
     void setNum(int num);
     void setType(std::string type);
     void setNom(std::string nom);
-    void setTps(int tps);
+    void setTps(float tps);
 };
